apps/backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Pillow>=10.0.0
Django==5.2
django-ninja==1.4.1
pydantic==2.11.4
pydantic_core==2.33.2
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
mysqlclient>=2.2.7
django-cors-headers==4.7.0
gunicorn==20.1.0
uvicorn[standard]==0.29.0
loguru==0.7.2
boto3>=1.28.0
djangorestframework==3.15.1
deep-translator==1.11.4
deepl==1.16.0
daphne==4.1.2
channels==4.1.0
channels_redis==4.2.0
packaging==24.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
pyotp==2.9.0
cryptography==42.0.5
